restapi/serializers.py
from rest_framework import serializers
from restapi.models import Book


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Book
        fields = ('id', 'title', 'author', 'isbn', 'price', 'description')
