requirements.txt
django==1.11.2
psycopg2==2.7.1
djangorestframework==3.6.3
django-filter
markdown
gunicorn
docker-py
dockerpty
docopt
isort==4.2.15
pylint==1.7.2
pylint-django